Мигрировал компоненты задач на TypeScript и переработал API
- Переименовал AddTask.vue в AddTaskModal.vue для соответствия нейминга
- Добавил типизацию TypeScript во все компоненты задач
- Обновил импорты с относительных путей на абсолютные (@/)
- Типизировал все реактивные переменные и параметры функций
- Добавил типы City и Nullable для корректной типизации
- Переработал вызовы API методов с использованием .execute()
- Обновил параметры методов для соответствия новой архитектуре
- Удалил неиспользуемые методы из TasksStore
- Убрал избыточный ref accounts из AccountsStore
- Обновил компоненты для работы с данными из API response
- Исправил типизацию событий в обработчиках
- Добавил автоматическое обновление списка задач после удаления

+++ frontend-vue/src/components/Tasks/Modals/DeleteAllTasksModal.vue
@@ -16,8 +16,11 @@
 
   const deleteTasks = () => {
     disable.value = true
-    tasksStore.deleteAllTasks(selectedTasksStatus, selectedAccountId)
-      .then(() => closeModal(modalId))
+    tasksStore.deleteAllTasks.execute({ status: selectedTasksStatus, accountId: selectedAccountId })
+      .then(() => {
+        closeModal(modalId)
+        tasksStore.fetchTasks.execute()
+      })
       .finally(() => disable.value = false)
   }
 </script>
